pyproject.toml
[tool.poetry]
name = "sam-envrironment"
version = "0.1.0"
description = "sam envrironment project"
authors = ["Nick Mazurkin <mazurkin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.12"

[tool.poetry.group.platform.dependencies]
setuptools = "80.3.1"

[tool.poetry.group.util.dependencies]
tqdm = "4.67.1"
argh = "0.29.4"

[tool.poetry.group.pytorch.dependencies]
torch = {url="https://download.pytorch.org/whl/cu130/torch-2.9.1%2Bcu130-cp312-cp312-manylinux_2_28_x86_64.whl"}
torchaudio = {url="https://download.pytorch.org/whl/cu130/torchaudio-2.9.1%2Bcu130-cp312-cp312-manylinux_2_28_x86_64.whl"}
torchvision = {url="https://download-r2.pytorch.org/whl/cu130/torchvision-0.24.1%2Bcu130-cp312-cp312-manylinux_2_28_x86_64.whl"}
torchcodec = "0.9.1"

[tool.poetry.group.huggingface.dependencies]
huggingface_hub = {version="0.36.0", extras=["cli"]}
transformers = {version="4.57.6", extras=["torch"]}

[tool.poetry.group.sam.dependencies]
sam-audio = {git="https://github.com/facebookresearch/sam-audio", rev="68b48d48fff1ad776d3afefbe634eb5f5d60ba7b"}

[tool.poetry.group.format.dependencies]
pyarrow = "22.0.0"
pyyaml = "6.0.3"

[tool.poetry.group.dev-lint.dependencies]
flake8 = "7.3.0"
pylint = "4.0.4"

[tool.poetry.group.dev-test.dependencies]
pytest = "8.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "slow: the tests takes a long time (deselect with '-m \"not long\"')",
]
